Pushed the complexity out of the delivery plugin
Based on feedback from Marcin, we should make the continuous delivery plugin cleaner :)
